Avoid possible UB in BigInt::from_s32
Computing -INT_MIN is undefined, cast to int64_t first.

+++ src/lib/math/bigint/bigint.cpp
@@ -43,7 +43,7 @@ BigInt BigInt::from_s32(int32_t n) {
    if(n >= 0) {
       return BigInt::from_u64(static_cast<uint64_t>(n));
    } else {
-      return -BigInt::from_u64(static_cast<uint64_t>(-n));
+      return -BigInt::from_u64(static_cast<uint64_t>(-static_cast<int64_t>(n)));
    }
 }
